+++ react-project-syncro/src/styles/Footer.css
