+++ react-project-syncro/src/styles/Sidebar.css
