+++ react-project-syncro/src/styles/SignUp.css
